migrate test_redismod to pytest
